+++ Basic/Strict-mode-ErrorRecord-formatting/Test-1.Out-String.ps1
@@ -1,6 +1,7 @@
 
 # v6-rc: fixed
 [PowerShell]::Create().AddScript({
+	$ErrorView = 'NormalView' #! v7
 	Set-StrictMode -Version 2
 	try {
 		throw 'Oops!'