+++ Cmdlets/Get-ItemPropertyValue/ErrorAction-ignored/.test.ps1
@@ -1,5 +1,5 @@
-
 if ($PSVersionTable.PSVersion.Major -le 4) {return task skip4}
+$ErrorView = 'NormalView' #! v7
 
 task Test-1.missing.item {
 	($r = try {.\Test-1.missing.item.ps1} catch {$_})
@@ -9,7 +9,7 @@ task Test-1.missing.item {
 }
 
 task Test-2.existing.item {
-	($r = exec {Invoke-PowerShell -NoProfile -Command .\Test-2.existing.item.ps1} | Out-String)
+	($r = exec {Invoke-PowerShell -NoProfile -Command '$ErrorView = ''NormalView''; .\Test-2.existing.item.ps1'} | Out-String)
 	assert ($r -clike '*FullyQualifiedErrorId*FullyQualifiedErrorId*invoked #1*FullyQualifiedErrorId*invoked #2*')
 }