+++ PowerShell.exe/Exit-code-0-with-File-script-error/.test.ps1
@@ -3,7 +3,12 @@ $Version = $PSVersionTable.PSVersion.Major
 
 task Test-Command-syntax {
 	($r = .\Test-Command-syntax.ps1 | Out-String)
-	assert ($r.TrimEnd() -clike '*MissingEndCurlyBrace*Exit code: 1')
+	if ($Version -ge 7) {
+		assert ($r.TrimEnd() -clike 'ParserError:*Exit code: 1*')
+	}
+	else {
+		assert ($r.TrimEnd() -clike '*MissingEndCurlyBrace*Exit code: 1')
+	}
 }
 
 task Test-Command-throw {
@@ -14,7 +19,10 @@ task Test-Command-throw {
 # fixed in v5
 task Test-File-syntax {
 	($r = .\Test-File-syntax.ps1 | Out-String)
-	if ($Version -ge 5) {
+	if ($Version -ge 7) {
+		assert ($r.TrimEnd() -clike 'ParserError:*Exit code: 1*')
+	}
+	elseif ($Version -ge 5) {
 		assert ($r.TrimEnd() -clike '*MissingEndCurlyBrace*Exit code: 1')
 	}
 	else {